App/Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.48
psycopg2-binary
python-dotenv==1.0.1
pydantic==2.7.1
alembic==1.13.1
